+++ src/main.ts
@@ -9,6 +9,7 @@ import { ConfigService } from "@nestjs/config";
 import { Logger } from "nestjs-pino";
 import fastifyRateLimit from "@fastify/rate-limit";
 import { major } from "semver";
+import * as yaml from "yaml";
 
 import { AppModule } from "./app.module";
 
@@ -73,13 +74,13 @@ code | condition
   const options = new DocumentBuilder();
 
   if (configService.get("api.development")) {
-    options.addServer(`http://localhost:${String(configService.get("api.port"))}`, "Development");
+    options.addServer(`http://localhost:${String(configService.get("api.port"))}`, "Development", {});
   }
 
   options
-    .addServer(`https://${apiDomain}`, "Production")
-    .addServer(`https://beta.${apiDomain}`, "Beta")
-    .addServer(`https://alpha.${apiDomain}`, "Alpha")
+    .addServer(`https://${apiDomain}`, "Production", {})
+    .addServer(`https://beta.${apiDomain}`, "Beta", {})
+    .addServer(`https://alpha.${apiDomain}`, "Alpha", {})
     .setTitle(`@open-sauced/api.opensauced.pizza`)
     .setDescription(markdownDescription)
     .setVersion(`1`)
@@ -91,6 +92,15 @@ code | condition
   const document = SwaggerModule.createDocument(app, options.build(), {
     operationIdFactory: (controllerKey: string, methodKey: string) => methodKey,
   });
+
+  try {
+    const yamlSwaggerDoc = yaml.stringify(document);
+
+    // write the yaml swagger doc to the root project directory anytime the server starts
+    await writeFile(path.resolve(process.cwd(), "swagger.yaml"), yamlSwaggerDoc, "utf8");
+  } catch (e) {
+    console.log(e);
+  }
 
   const customOptions: SwaggerCustomOptions = { swaggerOptions: { persistAuthorization: true } };